+++ src/lib/db/schema/index.ts
@@ -1,6 +1,7 @@
 export * from "./enums";
 export * from "./tenants";
 export * from "./users";
+export * from "./auth";
 export * from "./teams";
 export * from "./templates";
 export * from "./series";